requirements.txt
PySide6>=6.5.0
pystray
Pillow>=9.0.0
pywin32>=305; platform_system=="Windows"
customtkinter>=5.0.0
psutil>=5.9.0
requests>=2.28.0
openmeteo-requests
requests-cache retry-requests numpy pandas